added additonal styling to the product pages. The booking form needs to be configured but will be done by jorge
