+++ client/src/App.js
@@ -6,7 +6,7 @@ import Home from './pages/Home';
 import Search from './pages/Search';
 import { ThemeProvider } from './utils/themeContext';
 import Footer from './components/Footer';
-import ProductDetails from './components/ProductDetails';
+import Product from './pages/Product'
 
 const httpLink = createHttpLink({
   uri: '/graphql',
@@ -50,7 +50,7 @@ function App() {
                 path='/search'
                 element={<Search />}
               />
-                <Route path="/product/:id" element={<ProductDetails />} />
+                <Route path="/product/:id" element={<Product />} />
             </Routes>
             <Footer />
           </>
