+++ client/src/components/AppFeatures.js
@@ -4,7 +4,6 @@ import { Button } from 'primereact/button';
 import { Link } from 'react-router-dom';
 
 const AppFeatures = () => {
-
     
     return (
         <div className='app-feat'>
