+++ server/models/Service.js
@@ -36,6 +36,11 @@ const serviceSchema = new Schema(
       ref: 'Category',
       required: true
     },
+    user: {
+      type: Schema.Types.ObjectId,
+      ref: 'User',
+      required: true,
+    },
     options: [optionSchema],
     images: [imageSchema],
     location: {
